+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - Forced request of large MD tree or properties (introduced in previous release)
   is now downloaded to a local file instead of being open in a new document tab
   (vscode limits the document size to 50MB)
+- Fixed
+  - When switching to another environment, Design Automation tree is now updated properly
 
 ## [1.4.5] - 2019-11-20